+++ src/frontend/apps/impress/src/features/docs/doc-editor/api/useDocAITransform.tsx
@@ -7,7 +7,8 @@ export type AITransformActions =
   | 'prompt'
   | 'rephrase'
   | 'summarize'
-  | 'beautify';
+  | 'beautify'
+  | 'emojify';
 
 export type DocAITransform = {
   docId: string;